UBUNTU: [Packaging] Change per-kernel getabis into ESM mode
Replace the getabis script with a version which gets the personal URL(s)
from a config file. The variant we currently use for Trusty seems to
hard code the source name into the script. Trying to make this
automatically taken from the changelog, so we could use the same script
for all kernels.

Ignore: yes
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>